CriticalRate を int 型から float 型に変更

+++ project1/EnemyData.h
@@ -15,5 +15,5 @@ struct EnemyData
 	int Gold;
 	int EXP;
 	Element Element;
-	int CriticalRate;
+	float CriticalRate;
 };
